requirements.txt
blessed==1.17.5
click==7.0
dill==0.3.1.1
enlighten==1.5.2
ipdb==0.12.3
matplotlib==3.1.3
numpy==1.17.4
pytest==5.4.3
PyYAML==5.3.1
rich==1.0.0
recommonmark==0.6.0
scikit-learn==0.22.1
scipy==1.4.1
Sphinx==3.0.4
sphinx-rtd-theme==0.4.3
tqdm==4.43.0
